1. 改良选曲界面功能 2. 尝试修复win7下播放bga任会报错的问题 3. 将MeasureScale值渗透到判定中

+++ Assets/EZR/TimeLines.cs
@@ -10,7 +10,7 @@ namespace EZR
         public int MaxLines { get => 9; }
         public float BPM = 120f;
 
-        public int EndTick = 0;
+        public float EndTick = 0;
         public int BPMIndex = 0;
         public List<Pattern.BPM> BPMList = new List<Pattern.BPM>();
         public int[] LinesIndex;
