+++ Assets/PatternUtils/Pattern.cs
@@ -8,7 +8,7 @@ namespace PatternUtils
     public partial class Pattern
     {
         public static int TickPerMeasure { get => 192; }
-        public int EndTick = 0;
+        public float EndTick = 0;
         public class Sound
         {
             public string id = "";
@@ -19,7 +19,7 @@ namespace PatternUtils
 
         public class BPM
         {
-            public int position = 0;
+            public float position = 0;
             public float bpm = 0;
         }
 
@@ -27,11 +27,11 @@ namespace PatternUtils
 
         public class Note
         {
-            public int position = 0;
+            public float position = 0;
             public int id = 0;
             public float vol = 0;
             public float pan = 0;
-            public int length = 0;
+            public float length = 0;
         }
         public class Track
         {
